Added the following patch from the dev@openidenabled.com mailing list: http://lists.openidenabled.com/pipermail/dev/attachments/20090926/9e3e0f5b/attachment-0001.bin
Original Message:
pelle at stakeventures.com
Sat Sep 26 15:30:46 PDT 2009
darcs patch: Added support for the new OpenID OAuth extension.

This patch was in the form of a Darcs patch, not a normal patch.  So solve this, I applied it to the Darcs repository found on openidenabled, then created a new diff file between the original Darcs repo and the new one (with the patch applied) so that I could apply it to this git repo.

All hunks in patch succeeded.