Handle boolean value to fix signature issue
If an AX attribute is set to boolean(true) value it results in 
undefined method `gsub' for true:TrueClass error. 

If it's set to false then its replaced by empty string resulting in 
to signature mismatch on the consumer, because the server still computes 
false AX value.

Fix is to set val to empty string if it's nil and if it's
boolean type then convert it to string.