tests: set all fetch() params in test_redirect_limit
The default values of body and headers ought to be "nil", but in fact
they end up as the string "0" in test_redirect_limit.

This causes all sorts of problems. When body is defined as "0", we
follow the conditional that leads us to conn.request_post instead of
conn.request_get, and request_post crashes since it expects a key-value
string instead of simply "0". The unexpected string for the headers
variable wreaks havoc as well; when headers is "0", the hash value
assignment at the beginning of fetch() fails, because Ruby can't convert
the string to a hash.

I'm not sure why these values are "0" instead of nil, but setting them
explicitly during the call fixes the bug.